requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
cisco-anyconnect-cli==0.5
click==8.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.12
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
idna==3.3
importlib-metadata==4.11.1
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
keepasshttp==1.4.1
MarkupSafe==2.1.0
openapi-codec==1.3.2
ordereddict==1.1
packaging==21.3
pycryptodome==3.14.1
PyJWT==1.7.1
PyMySQL==1.0.2
pyparsing==3.0.7
pytz==2021.3
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.8
zipp==3.7.0
